Fixed lint errors
