+++ package.json
@@ -8,7 +8,8 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
-
+    "eslint-config-airbnb": "^18.2.0",
+    "eslint-plugin-jest": "^24.0.0",
     "firebase": "^7.19.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
@@ -19,9 +20,7 @@
     "react-scripts": "^3.4.3",
     "react-typing-effect": "^2.0.3",
     "reactstrap": "^8.5.1",
-    "typescript": "^4.0.2",
-    "eslint-config-airbnb": "^18.2.0",
-    "eslint-plugin-jest": "^24.0.0"
+    "typescript": "^4.0.2"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -34,6 +33,7 @@
     "extends": [
       "airbnb"
     ],
+    "parser": "babel-eslint",
     "plugins": [
       "jest"
     ],
