+++ src/App.jsx
@@ -1,27 +1,25 @@
-import React, { Component } from "react"
-import "./App.css"
-import Header from './components/Header';
-import Footer from './components/Footer';
-
+import React, {} from 'react';
+import './App.css';
 import {
   BrowserRouter as Router,
   Route,
-} from "react-router-dom";
+} from 'react-router-dom';
+
+import Header from './components/Header';
+import Footer from './components/Footer';
 
 import Home from './pages/Home';
 
-class App extends Component {
-  render() {
-    return (
-      < Router >
-        <Header />
-        <Route>
-          <Home />
-        </Route>
-        <Footer />
-      </Router>
-    );
-  }
+function App() {
+  return (
+    <Router>
+      <Header />
+      <Route>
+        <Home />
+      </Route>
+      <Footer />
+    </Router>
+  );
 }
 
 export default App;
