--- src/components/Footer.js
@@ -1,12 +1,0 @@
-import React, { Component } from 'react'
-
-class Footer extends Component {
-    render() {
-        return (
-            <div class="footer">
-                <p></p>
-            </div>
-        )
-    }
-}
-export default Footer
