+++ src/components/Footer.jsx
@@ -1,0 +1,10 @@
+import React, { } from 'react';
+
+function Footer() {
+  return (
+    <div className="footer">
+      <p />
+    </div>
+  );
+}
+export default Footer;
