--- src/components/Header.js
@@ -1,13 +1,0 @@
-import React, { Component } from 'react'
-import LoginMenu from '../components/LoginMenu';
-
-class Header extends Component {
-   render() {
-      return (
-         <div class="header">
-            <LoginMenu />
-         </div>
-      )
-   }
-}
-export default Header
