+++ src/components/Header.jsx
@@ -1,0 +1,11 @@
+import React, { } from 'react';
+import LoginMenu from './LoginMenu';
+
+function Header() {
+  return (
+    <div className="header">
+      <LoginMenu />
+    </div>
+  );
+}
+export default Header;
