+++ src/pages/Home.jsx
@@ -1,0 +1,13 @@
+import React, { } from 'react';
+
+function Home() {
+  return (
+    <div className="App">
+      <h4>Hi there,</h4>
+      <h1>Find Scholarships Today</h1>
+
+    </div>
+  );
+}
+
+export default Home;
